requirements.txt
numpy
pandas
matplotlib
scikit-learn
kagglehub
PyYAML
dvc
nltk
logger